+++ packages/radix-ui-themes/src/theme-options.tsx
@@ -16,6 +16,7 @@ const appearances = ['inherit', 'light', 'dark'] as const;
 const accentScales = [...radixColorScales, 'gray'] as const;
 const grayScales = [...radixGrayScales, 'auto'] as const;
 const backgroundColors = ['auto', 'gray'] as const;
+const panelBackgrounds = ['solid', 'translucent'] as const;
 const textColors = ['auto', 'accent'] as const;
 const radii = ['none', 'small', 'medium', 'large', 'full'] as const;
 const scalings = ['90%', '95%', '100%', '105%', '110%'] as const;
@@ -26,6 +27,7 @@ const themePropDefs = {
   grayScale: { type: 'enum', values: grayScales, default: 'auto' },
   backgroundColor: { type: 'enum', values: backgroundColors, default: 'auto' },
   applyBackgroundColor: { type: 'boolean', default: true },
+  panelBackground: { type: 'enum', values: panelBackgrounds, default: 'translucent' },
   textColor: { type: 'enum', values: textColors, default: 'auto' },
   radius: { type: 'enum', values: radii, default: 'medium' },
   scaling: { type: 'enum', values: scalings, default: '100%' },
@@ -35,6 +37,7 @@ const themePropDefs = {
   grayScale: PropDef<(typeof grayScales)[number]>;
   backgroundColor: PropDef<(typeof backgroundColors)[number]>;
   applyBackgroundColor: PropDef<boolean>;
+  panelBackground: PropDef<(typeof panelBackgrounds)[number]>;
   textColor: PropDef<(typeof textColors)[number]>;
   radius: PropDef<(typeof radii)[number]>;
   scaling: PropDef<(typeof scalings)[number]>;
@@ -46,6 +49,7 @@ type ThemeAppearance = NonNullable<ThemeProps['appearance']>;
 type ThemeAccentScale = NonNullable<ThemeProps['accentScale']>;
 type ThemeGrayScale = NonNullable<ThemeProps['grayScale']>;
 type ThemeBackgroundColor = NonNullable<ThemeProps['backgroundColor']>;
+type ThemePanelBackground = NonNullable<ThemeProps['panelBackground']>;
 type ThemeTextColor = NonNullable<ThemeProps['textColor']>;
 type ThemeRadius = NonNullable<ThemeProps['radius']>;
 type ThemeScaling = NonNullable<ThemeProps['scaling']>;
@@ -55,6 +59,7 @@ type ThemeOptions = {
   accentScale: ThemeAccentScale;
   grayScale: ThemeGrayScale;
   backgroundColor: ThemeBackgroundColor;
+  panelBackground: ThemePanelBackground;
   textColor: ThemeTextColor;
   radius: ThemeRadius;
   scaling: ThemeScaling;